+++ packages/i18n/src/define.ts
@@ -1,4 +1,4 @@
-import { type LocaleMessages } from './locale'
+import type { LocaleMessages } from './locale'
 
 export const defineLocale = (locale: LocaleMessages) => {
   return locale